fix(sonar): clear S7632 prose-token and S3776 describe rollup
- _audit.py: rewrite the prose comment above the S7504 suppressions
  so it no longer contains the literal NOSONAR token; Sonar's S7632
  was interpreting the explanatory mention as an orphan directive.

- describe.py + _describe_helpers.py: hoist the post-validation
  orchestration (client open, metadata + table + choice + docs
  fetches, field filtering, response assembly) into a new
  _describe_impl helper. Brings the inner describe closure and the
  register_tools rollup under Sonar's S3776 threshold of 15.

All 772 tests pass; behavior, warning order, validation order,
and response shape unchanged.
