+++ src/servicenow_mcp/tools/describe.py
@@ -11,25 +11,20 @@ a single source of truth.
 """
 
 import logging
-from typing import Any
 
 from mcp.server.fastmcp import FastMCP
 
 from servicenow_mcp.auth import BasicAuthProvider
 from servicenow_mcp.choices import ChoiceRegistry
-from servicenow_mcp.client import ServiceNowClient
 from servicenow_mcp.config import Settings
 from servicenow_mcp.decorators import tool_handler
 from servicenow_mcp.policy import check_table_access
 from servicenow_mcp.tools._describe_helpers import (
-    _build_slim_field_list,
-    _build_verbose_field_list,
-    _fetch_choice_counts,
-    _fetch_documentation,
+    _describe_impl,
     _parse_fields_filter,
 )
 from servicenow_mcp.tools._dictionary import DictionaryRegistry, ScriptField
-from servicenow_mcp.utils import ServiceNowQuery, format_response, validate_identifier
+from servicenow_mcp.utils import format_response, validate_identifier
 
 
 logger = logging.getLogger(__name__)
@@ -37,26 +32,6 @@ logger = logging.getLogger(__name__)
 TOOL_NAMES: list[str] = ["describe"]
 
 _VALID_DESCRIBE_ACTIONS: frozenset[str] = frozenset({"list_script_fields"})
-
-
-def _apply_fields_filter(
-    field_list: list[dict[str, Any]],
-    requested_fields: list[str],
-    warnings: list[str],
-) -> list[dict[str, Any]]:
-    """Restrict ``field_list`` to ``requested_fields`` and warn on unknown names.
-
-    Preserves left-to-right order of ``requested_fields`` when collecting the
-    ``unknown`` list so the warning text is stable across runs. The returned
-    list keeps the original ``field_list`` ordering for matched fields.
-    """
-    wanted = set(requested_fields)
-    present = {str(f.get("name") or f.get("element") or "") for f in field_list}
-    unknown = [name for name in requested_fields if name not in present]
-    filtered = [f for f in field_list if str(f.get("name") or f.get("element") or "") in wanted]
-    if unknown:
-        warnings.append(f"Unknown field(s): {','.join(unknown)}")
-    return filtered
 
 
 def _script_field_summary(fields: list[ScriptField]) -> list[dict[str, object]]:
@@ -180,45 +155,14 @@ def register_tools(
         for name in requested_fields:
             validate_identifier(name)
 
-        warnings: list[str] = []
-
-        async with ServiceNowClient(settings, auth_provider) as client:
-            metadata = await client.get_metadata(table)
-
-            # Fetch table-level metadata from sys_db_object
-            table_meta = await client.query_records(
-                "sys_db_object",
-                ServiceNowQuery().equals("name", table).build(),
-                fields=["label", "super_class", "is_extendable", "number_ref", "sys_id"],
-                limit=1,
-            )
-            table_info = table_meta.get("records", [{}])[0] if table_meta.get("records") else {}
-
-            # Batched sys_choice fetch keyed by field; non-fatal failures so a
-            # slim describe still works in restricted instances.
-            choice_counts = await _fetch_choice_counts(client, table, warnings)
-
-            # Optional sys_documentation fetch (off by default; help text is huge).
-            docs: dict[str, dict[str, Any]] = {}
-            if include_docs:
-                docs = await _fetch_documentation(client, table, warnings)
-
-        field_list = (
-            _build_verbose_field_list(metadata, choice_counts)
-            if verbose
-            else _build_slim_field_list(metadata, choice_counts)
+        data, warnings = await _describe_impl(
+            table,
+            verbose=verbose,
+            include_docs=include_docs,
+            requested_fields=requested_fields,
+            settings=settings,
+            auth_provider=auth_provider,
         )
-
-        if requested_fields:
-            field_list = _apply_fields_filter(field_list, requested_fields, warnings)
-
-        data: dict[str, Any] = {
-            "table": table_info,
-            "fields": field_list,
-            "field_count": len(field_list),
-        }
-        if include_docs:
-            data["documentation"] = docs
 
         return format_response(
             data=data,
